+++ frontend/package.json
@@ -13,6 +13,7 @@
     "axios": "^1.6.8",
     "framer-motion": "^11.2.4",
     "jwt-decode": "^4.0.0",
+    "lodash": "^4.17.21",
     "partysocket": "^1.0.1",
     "react": "^18.2.0",
     "react-calendar-heatmap": "^1.9.0",
