+++ frontend/public/styles/Training.css
