+++ frontend/src/components/ChatComponent.jsx
