+++ frontend/src/components/Line.jsx
